requirements.txt
Flask>=2.0
python-telegram-bot>=20.0
google-cloud-vision>=3.0.0
pytesseract>=0.3.10
Pillow>=9.0
openai>=1.0.0
python-dotenv>=1.0
pytest>=7.0
sqlalchemy>=1.4
requests>=2.28